bug(#3760): Synchronized StrictXmir

+++ eo-parser/src/main/java/org/eolang/parser/StrictXmir.java
@@ -205,24 +205,29 @@ public final class StrictXmir implements XML {
     private static File fetch(final String uri, final Path path, final Path tmp) {
         final File ret;
         if (StrictXmir.MINE.equals(uri)) {
-            if (path.toFile().getParentFile().mkdirs()) {
-                Logger.debug(StrictXmir.class, "Directory for %[file]s created", path);
+            synchronized (tmp) {
+                if (!path.toFile().exists()) {
+                    if (path.toFile().getParentFile().mkdirs()) {
+                        Logger.debug(StrictXmir.class, "Directory for %[file]s created", path);
+                    }
+                    try {
+                        System.out.println("HERE");
+                        Files.write(
+                            path,
+                            new IoCheckedBytes(
+                                new BytesOf(new ResourceOf("XMIR.xsd"))
+                            ).asBytes()
+                        );
+                        Logger.debug(StrictXmir.class, "XSD copied to %[file]s", path);
+                    } catch (final IOException ex) {
+                        throw new IllegalArgumentException(
+                            String.format("Failed to save %s to %s", uri, path),
+                            ex
+                        );
+                    }
+                }
+                ret = path.toFile();
             }
-            try {
-                Files.write(
-                    path,
-                    new IoCheckedBytes(
-                        new BytesOf(new ResourceOf("XMIR.xsd"))
-                    ).asBytes()
-                );
-                Logger.debug(StrictXmir.class, "XSD copied to %[file]s", path);
-            } catch (final IOException ex) {
-                throw new IllegalArgumentException(
-                    String.format("Failed to save %s to %s", uri, path),
-                    ex
-                );
-            }
-            ret = path.toFile();
         } else {
             ret = StrictXmir.download(uri, path, tmp);
         }
